fix: connections has a cap of 100 instead of size
Previously, I set the size of the array to 100, which pre-populates the
array with nils. The capacity was the one that should have been set.

Another issue was that using range over the array is which is being
modified. range copies the array and iterates over that. If the array
was modified, then you'll get NPEs when trying to access an element by
index. Instead of that, iterating by checking size of array is better